Renamed VISPR to ViRelAy and improved README
Various filenames have been changed to virelay. All descriptions, titles
etc. were changed in backend and frontend. Ability to specify project
paths directly with gunicorn was added. Node-modules were updated and
the front-end static files were updated. The README was improved and the
screenshot was updated to the current version.