Fixed the overlay image mode
When there are no positive attributions in an attribution, then the
overlay image mode will only display the heatmap and not the heatmap
superimposed onto the input image. This is because of the way that the
negtive attribution mask, which is used to superimpose the negative part
of the attributions onto the input image, is created. It is created by
clipping the attributions multiplied by -1 to a range of 0 and the
maximum value of the attributions multiplied by -1. But since there are
no negative attributions, the maximum value of the attributions
multiplied by -1 is negative. Thus, the maximum of the clipping range is
smaller than the minimum and the resulting mask has negative values in
it. Since this value is very small, the actual alpha mask, which is
created from the mask, is almost pure white, thus mixing in the heatmap
with almost 100% and leaving nothing of the actual image.