Updated the Python dependencies
Updated the dependencies in the setup.py file to the their respective
latest versions. Previously, the versions of the "pytest" and
"pytest-cov" packages were not pinned to specific versions, but now they
are pinned to the latest versions. Futhermore, the package versions are
now restricted to be below the next major version, so that we do not
pull in any breaking changes in the future.

Python 3.7 has already reached its end of life June 27, 2023. Also,
Python 3.8 will reach its end of life in October 2024, which is
imminent. Therefore, both versions were removed from the supported
Python versions in the setup.py file, the tox configuration, and in the
GitHub Actions workflow file. In exchange, Python 3.10, 3.11, and 3.12
were added to the supported Python versions. This means that now Python
3.9, 3.10, 3.11, and 3.12 are the supported Python versions for the
project. The documentation was also updated to reflect these changes.

The updated dependencies and Python versions required some minor changes
in the project:

- The configuration Python file for building the documentation
  "docs/source/conf.py" was using the "pkg_resources" module, which has
  been deprecated and was removed in Python 3.12. The
  "importlib.metadata" module is now used as a replacement.
- In a previous version, Flask used "application/javascript" as the MIME
  type for JavaScript files. However, the correct MIME type is
  "text/javascript", which is now also used by Flask. For this reason,
  the unit test for it was updated.
- The server previously also used the "pkg_resources" module to serve
  the static frontend files from the package directory. Since the module
  is now deprecated and was removed in Python 3.12, it was replaced with
  the "importlib.resources" module.

In the tox configuration file, no version numbers were specified PyTest,
PyLint, and Flake8, which causes the tests to fail, as tox is installing
the latest versions which have changed some of the configuration
options. Therefore, the versions of PyTest, PyLint, and Flake8 were
fixed to the versions that were current when the tests were last
updated. This is just a temporary fix until the unit tests and the
linters are updated to work with the latest versions of PyTest, PyLint,
and Flake8.

Additionally, in the GitHub Actions workflow configuration file, the
"jobs.docs.strategy.fail-fast" option was previously used to prevent the
workflow from stopping if the documentation build failed. However, this
option is only supported for matrix strategies. Since the "docs" job
does not use a matrix strategy, the "jobs.docs.strategy.fail-fast"
option was replaced with the "jobs.docs.continue-on-error" option.