Improved the Python Linting
The Flake8 linter was replaced by PyCodeStyle and PyDocLint, and MyPy
was added as a static type checker. All four checkers are now included
in the setup.py as extra dependencies under the name "linting", so that
developers can install them if they want to use them locally, e.g., as a
Visual Studio Code integration.

New configurations for PyCodeStyle, PyDocLint, and MyPy were added and
the configuration file for PyLint was renamed to ".pylintrc", to match
the naming convention of the other configuration files, and was updated
to match the latest version of PyLint.

The PyLint and Flake8 test environments were removed from the tox.ini
file and the new test environment "linting" was added, which runs all
linters and the static type checker. The maximum line length was
increased from 120 to 150, which makes it now easier to break some
longer lines.

The Flake8 and PyLint jobs were removed from the GitHub Actions Workflow
configuration file, and the new job "lint-and-type-check" was added,
which runs the tox test environment "linting" to run the linters and the
static type checker. Also, the GitHub Actions Workflow configuration
file was cleaned up and documented.

The documentation was updated to reflect the changes in the linting
process and to explain how to run the linters and the static type
checker locally.

As it turns out, Python 3.9 is missing the union operator "|" for type
hints. Although this operator is not strictly necessary, it makes the
type hinting for MyPy much easier to read. Also, the latest NumPy
version does not support Python 3.9 anymore. Since ViRelAy is not a
library used by other projects and since most operating systems already
support Python 3.10, the minimum Python version was updated to 3.10 to
make the development process easier. Python 3.9 was removed from the tox
configuration, the GitHub Actions Workflow configuration, and from the
documentation. Also, the NumPy version that was previously fixed to
2.0.2, which is the last version with Python 3.9 support, was updated to
the latest version, which is 2.1.1.

The Ubuntu and Python versions used by "Read the Docs" to build the
documentation were also updated to the latest versions, i.e., Ubuntu
24.04 and Python 3.12.

The entire code base was linted and type-checked, and all issues found
by the linters and the type checker were fixed. Thanks to the new
linters and the new type checker, several obscure bugs were found that
would have gone unnoticed otherwise:

1.  The automatic reload of the Flask server was not working correctly,
    because instead of using the property "debug", the incorrect property
    "auto_reload" was used.
2.  The ViRelAy modules were imported using the relative import syntax
    instead of the absolute import syntax, which may cause problems
    because sometimes it is not clear if an import is relative or not,
    relative imports can be ambiguous, they are brittle and may break
    when a module is moved, and since Python 3, implicit relative imports
    are not allowed anymore.
3.  The "get_label..." methods of the LabelMap class returned the
    human-readable names of the labels instead of the labels themselves,
    which was not obvious from the method names. In some parts of the
    code, the list of human-readable names was used instead of a list of
    labels, which was incorrect. Now, the LabelMap has a new set of
    methods for retrieving the labels, and the old methods were renamed
    to make their purpose clearer. All instances where the methods are
    used were checked and corrected if necessary.
4.  In multiple instances, variables were reused, for example, variables
    that were used to store a NumPy array of an image were reused when
    creating a Pillow Image from them. This lead to some errors, where
    a NumPy array was expected, but a Pillow Image was passed instead,
    and vice versa.
5.  In many cases, the documentation for methods was either missing,
    incomplete, or incorrect. The documentation was updated to reflect
    the actual behavior of the methods.
6.  In the tests, the names of the fixtures were reused and may therefore
    have accidentally overwritten each other. The names of the variables
    and parameters were changed to be unique.
7.  The unit tests were missing a __init__.py file and were thus not a
    proper Python package. This was fixed by adding the missing file.
8.  The __init__.py file of the ViRelAy package did not contain a
    docstring, which was added.
9.  The fixtures for the unit tests are nested, which means they
    reference each other. Since the parameter names must match the names
    of the fixtures and since the fixtures were all in the same file, the
    parameters were shadowing the fixtures. This was fixed by renaming
    the fixture functions to "get_..._fixture" and adding a name to the
    fixture attribute.
10. Some unit tests were incorrectly comparing "Label" objects with
    strings, which is now fixed.
11. The Sphinx build configuration script was previously updated to use
    importlib.metadata instead of the deprecated pkg_resources package,
    which was removed in Python 3.12. Unfortunately, the new
    implementation was using the importlib.metadata.requires function to
    replace the pkg_resources.require function, which was incorrect, as
    they are not equivalent. This was fixed by entirely removing the
    importlib.metadata package and simply using the module.__file__
    attribute to get the path to the ViRelAy package. This requires some
    more path manipulation, but it is much easier to understand and
    maintain.