Fixed PyCodeStyle
As it turned out, the semantics of the select option must have changed.
Instead of selecting a warning or error in addition to all the ones that
are selected by default, it only selects the ones that are explicitly
listed. Therefore, I removed the option to select a single warning,
which seems to have been removed anyway. Now, PyCodeStyle will correctly
report errors. All errors reported by PyCodeStyle have been fixed.