Removed the "LICENSE" file
Previously, there were two license files: "COPYING", which contained the
AGPL 3.0 license text, and "LICENSE", which contained a note about where
to find the license of the project and any third-party licenses. This
was done, because GPL prefers the file name "COPYING", but back then
GitHub did not support that file name. Now, GitHub also supports
"COPYING". As the information about where to find the project license
and the third-party licenses is also contained in the read me, the
"LICENSE" file was removed.