Added a way to run tox in Docker
Added a Dockerfile for a Docker image that contains tox and all
supported Python versions. This makes it easy to run the unit tests with
all supported Python versions, without having to install multiple Python
versions on your system. A container can be run using a convenience
script that will automatically build the Docker image, if it is not
already locally available, an run tox inside of it. The convenience
script can be used as a drop-in replacement for tox.