Added more type hints to the codebase
In order to improve MyPy's ability to reason about the types and
therefore find bugs, many type hints were added throughout the code.
Types were introduced to make the dictionaries that are loaded from or
written to YAML/JSON files strongly typed. Also, types were added for
strings that have a specific set of possible values.

For the literal types for strings with a specific set of possible
values, type aliases were created, so that the same type can be reused
in multiple places. This makes it easier to change the set of possible
values in the future, as the change only needs to be made in one place.
This, however, also made it necessary to change the code for generating
the GitHub URLs for the source code links, because literal types are no
real types, but instances of a class. The regular code for finding the
file name and the line numbers of the source code corresponding to
modules, classes, methods and functions cannot be used. Instead, an
alternative approach was implemented that assumes that the type alias
for the literal type is located in the same file as the module it
belongs to. This is a reasonable assumption and works for all current
uses of literal types. To find the line number of the type alias, the
source code of the module file is loaded and parsed using a regular
expression that matches the type alias definition. This approach is not
as robust as using the "inspect" module, but it works well in practice.

Finally, the year of the copyright notice in the Sphinx build
configuration is now always set to the current year, which means we do
not have to remember to update it every year.