Separate Jobs for the Linters & Type Checker
The GitHub Actions Workflow job for the linters and the static type
checker was split up into four separate jobs, one for each linter or
static type checker. Since the jobs invoke corresponding tox
environments, the tox environment for the linters and the static type
checker was also split up into separate environments for the linters and
static type checker. The documentation was updated to reflect the
changes. Also, some comments were added to the tox file to make the
different sections and settings more clear.