Converted the backend project to a uv project
There is a new Python package and project manager called uv. It is
written in Rust, which makes it blazing fast. First of all, the old way
of packaging Python projects using setuptools and "setup.py" are dated
and lack certain features that make it increasingly complicated to
develop the project.

In this commit, the backend REST API project was moved into a "source"
directory and the frontend project was moved from the backend REST API
project into the source directory as well. This was, for example, one
problem, which made using setuptools impossible, because it does not
allow to include data file (in this case the files of the built
frontend) that is not located inside the directory of the Python
package. The uv project manager and its build system hatchling make it
extremely easy to achieve this.

Also, uv manages virtual environments and allows us to install multiple
different versions of Python, which means that we no longer have to
install multiple Python versions on the OS-level or use the custom tox
Docker image. For this reason, the Docker tox project was also removed.

The tox configuration and the GitHub Actions workflow were adapted to
use uv, which makes it much easier to run them.

The documentation was updated to include information about how to use
uv with the project and how to run ViRelAy, the unit tests, perform the
linting and run the static type analysis. Furthermore, the
documentation was generally revised and edited. Some parts were
improved and some minor mistakes were fixed.

The following tasks were also performed:

- The example scripts and the inline scripts in the documentation were
  linted and their types were statically checked. All issues that were
  identified have been fixed.
- The way the frontend files were loaded in the backend REST API was
  updated to include the new location they are stored.
- Some minor code formatting problems were fixed.
- The "tox.ini" file was now also moved to the "tests/config"
  directory.