Updated the frontend project
The frontend of ViRelAy is an Angular SPA application, which uses
Clarity as UI library. The old project was still running on Angular 13
and Clarity 13, which have been superseded by Angular 18 and Clarity 17.
Also, all of its dependencies were outdated. For this reason, the
project was migrated to the latest Angular and Clarity versions and all
of its dependencies were updated to their respective latest versions.
Futhermore, the project was completely refactored and updated to use the
latest technologies and patterns. This includes the following tasks:

- The project was previously nested in another "src" directory, which is
  weird, because the project is not at the top-level of the repository,
  but in a source directory itself. For this reason, the files were
  moved out of the "src" directory and one level up.
- The "main.ts" and "index.html" files were moved into the app
  directory.
- The Angular and TypeScript configurations were completely overhauled
  and commented.
- A favicon with support for all browsers, light and dark themes, iOS,
  and Android was created.
- The module paths are now remapped to "@app", "@components", "@pages",
  and "@services", which makes it nicer to import them, because no
  relative paths are needed anymore.
- The "components" and "modules" directories were moved out of the app
  directory into the top-level directory of the project and "modules"
  was renamed to "pages", because it only contains components that are
  pages.
- The components are now standalone, which means that they are no longer
  included in any modules and can be loaded individually.
- Proper error handling for the bootstrap process was integrated.