Added multiple linters for the frontend
Added ESLint, a JavaScript and TypeScript linter, Stylelint, a CSS and
Sass linter, and HTML-Validate, an HTML linter. All issues that were
detected by the linters was fixed. The documentation was updated to
include instructions on how to run the linters. The GitHub Actions
workflow configuration was adapted to run the linters.