Made the selected attributions pane responsive
Added two breakpoints at 1080px and 720px window height, which adapt the
height of the selected attributions pane (from 250px to 200px and 150px
respectively). This prevents the selected attributions pane from taking
up too much space when the browser window is small.

Also, increased the maximum zoom level to 10 and decreased the minimum
zoom level to 0.1. When the embedding viewer is too small, then the new
minimum zoom level allows users to fit more embedding vectors into the
viewport, and the new maximum zoom level allows users to zoom in deeper
when the window is large.