Updated the dependencies
Updated the dependencies of the backend, the frontend, and the GitHub
Actions workflow to the latest versions. For the backend, the supported
Python versions were updated from 3.13.0, 3.12.7, 3.11.10, and 3.10.15
to 3.13.2, 3.12.9, 3.11.11, and 3.10.16 respectively. Also, the
dependencies of the backend were updated to the latest versions
available. For the frontend the dependencies, most notably the Angular
framework but also the dependencies of the linters, were also updated to
the latest versions available. Finally, the actions used in the GitHub
Actions workflow were updated to the latest versions available.

By updating the linters for the backend and the frontend, some
configuration problems, as well as some new linting errors were
introduced. These configuration issues and errors were fixed in the
codebase.

Previously, before running the unit tests of the backend, the frontend
needed to be built manually. This was required, because the unit tests
test if the frontend could be served through the backend, for which the
frontend build artifacts are required. This was a cumbersome process and
it required a warning in the documentation, which was not ideal,
especially for new contributors. This process was automated by adding
the building of the frontend to the tox configuration. Now, the frontend
is built automatically before running the unit tests of the backend.
This way, the unit tests can be run without any manual steps and
therefore, the warning in the documentation was removed.

Since this is the first update of ViRelAy for the new release, the
version number was updated from 0.5.0 to 0.6.0. Also, the changelog was
updated to reflect the changes made in this commit.

Finally, some minor spelling mistakes in the documentation were also
fixed.