Improved the Image Sampling in the Frontend
Previously, the previews of the samples and heatmaps were sampled either
using nearest neighbor or bilinear sampling. This commit introduces a
new option in the UI, which allows users to select the sampling method
for the images. By default, the option is set to auto, which will select
a method based on sizes of the sample image and the HTML element. The
user can, however, override this option and select either nearest
neighbor or bilinear sampling. This will make it easier to inspect the
samples and especially the heatmaps, which are often very small images
and therefore get very blurry when sampled using bilinear sampling.

The documentation was updated to include this new feature. As the
feature caused some changes in the UI, the screenshots in the
documentation were updated as well.