Fixed the Sample and Heatmap Alternative Text
Fixed the alternative text of the sample and heatmap images in the
sample viewer. Previously, the text was "[Object object]", which is the
default string that is returned when the "toString" method of an object
is called. The alternative text of the sample images and heatmaps is
generated from their sample labels. The problem was, that the labels
were not strings, but objects containing the label index, the WordNetId,
and the human-readable name of the label.