Moved Python Project File Into Root Directory
During the build of the documentation on ReadTheDocs, the ViRelAy
package needs to be available, as it is referenced in the configuration
file of the documentation. Unfortunately, ReadTheDocs requires the
"pyproject.toml" file to be in the root directory of the repository.
Therefore, the "pyproject.toml" file has been moved from the
"source/backend" directory to the root directory of the repository.
Also, the "uv.lock", ".python-versions" and the "tox.ini" files have
been moved to the root directory. All references to the these files, as
well as relative paths in the code, the tox configuration file, the
GitHub Actions workflow files, and the documentation were updated
accordingly. Some necessary changes where made to the "pyproject.toml"
file to ensure that the package can still be built and installed
correctly.