Updated the ViRelAy Image Used in the PyPI Read Me
The read me of the repository is also used as the project description,
which is displayed on the PyPI project page. The read me contains two
screenshots of ViRelAy's UI: One of the light-mode UI and one of the
dark-mode UI. The read me uses a GitHub feature that allows us to
display the light-mode screenshot when the user uses GitHub's
light-mode, and the dark-mode screenshot when the user uses GitHub's
dark-mode. PyPI does not support this feature and would display both
screenshots. Therefore, the dark-mode screenshot was removed from the
read me before it was included in the package using the "Hatch Fancy
PyPI Readme" plugin. To put emphasis on the fact that ViRelAy supports
both light-mode and dark-mode, both screenshots are now replaced with a
single image that contains a split-screen montage of ViRelAy's
light-mode and dark-mode UIs.