Cleaned Up the Unit Tests & Linters
Restructured the "tests" directory to better organize unit tests and
linter configurations. First of all, the configurations for the linters
have been moved to a new directory called "tests/linters". Converted the
configuration for the HTMLValidate linter to a Node.js package, so that
it is easier to use and maintain. The configurations for PyTest and
Coverage.py were extracted from the tox configuration file and placed
into their own files in the "tests/unit_tests" directory. All references
to the old configuration files in the "tests/config" directory have been
updated to point to the new locations.