Fixed the Usage of uv in the tox Configuration
A major problem in the tox configuration was fixed. Previously, tox-uv
was always using the system Python version instead of the Python
versions installed via uv. This meant that the Python versions specified
in the environment names and the "base_python" option were ignored. This
meant that the unit tests were not run with the correct Python version
and the environments that used the "base_python" option were not created
with the correct Python version. Previously, a fix was attempted by
setting the "uv_python_preference" to "only-managed", but this was not
the entire solution. What finally fixed the issue was to always use the
commands directly instead of running them through the "uv run" command.

Also, some problems with the MyPy configuration now emerged: MyPy could
not find the type information for PyTest and Sphinx. This was fixed by
adding the "pytest-mypy" package and ignoring the "sphinx" package in
the MyPy configuration.