Updated the Dependencies
Updated the versions of Python that are specified in the
".python-versions" file and the "pyproject.toml" file to the latest
patch versions of Python 3.11, 3.12, and 3.13:

- Python 3.13.2 → 3.13.5,
- Python 3.12.9 → 3.12.11, and
- Python 3.11.11 → 3.11.13

Updated the dependencies of the backend REST API and the frontend to
their respective latest versions. Also updated the versions of the
actions, the Node.js version used to build the frontend and run the
linters, and the uv version that are used in the GitHub Actions
workflows to their latest versions.

Also, the following changes were made:

- The "uv.lock" file is now disabled when synchronizing the package
  versions during deployment. This causes uv to assume that the lock
  file will not change and will raise an error if it does. This way, the
  deployment will fail if the lock file is not up-to-date and not
  silently ignore the changes and deploy anyway.
- During the deployment, the tags are now explicitly fetched from the
  remote repository, so that the build system is able to determine the
  latest version of the package, which is used to set the version of the
  package being build and deployed.
- Fixed an issue found by the newer version of ESLint, where a numeric
  enumeration value was cast to a number, which is not necessary.
- Removed the deprecated mode parameter from the "fromArray" method
  of the PIL Image class in the "image_processing.py" file, which was
  causing a warning during the build process.